Split parser in parsing modules
